requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pyjwt==2.10.1
bcrypt==4.2.1
cryptography==43.0.1
email-validator==2.2.0
httpx==0.28.1
python-keycloak==4.3.0
supabase==2.10.0
